LayoutTests/http/tests/loading/preload-img-srcset-sizes.html

<!DOCTYPE html>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<script src="http://127.0.0.1:8000/resources/slow-script.pl?delay=1000"></script>
<script src="http://127.0.0.1:8000/resources/checkPreload.js"></script>
<body>
This test makes sure that the Preload scanner preloads only the right image resource.
<br/>
<script>
checkForPreload("resources/preload-test.jpg?1", true);
checkForPreload("resources/preload-test.jpg?2", false);
</script>
<img sizes="50vw" srcset="resources/preload-test.jpg?1 640w, resources/preload-test.jpg?2 1000w">
</body>
