LayoutTests/http/tests/loading/preload-img-srcset-sizes-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This test makes sure that the Preload scanner preloads only the right image resource.
PASS PASS
